clean up settings page
